frontend/backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
groq>=0.4.0
python-dotenv>=1.0.0
firecrawl-py>=1.0.0
